the Users and Signature classes have been moved to the domain/entities package
