+++ src/main/java/br/com/pipocaagil/apipipocaagil/config/ModelMapperConfig.java
@@ -1,6 +1,6 @@
 package br.com.pipocaagil.apipipocaagil.config;
 
-import br.com.pipocaagil.apipipocaagil.domain.Users;
+import br.com.pipocaagil.apipipocaagil.domain.entities.Users;
 import br.com.pipocaagil.apipipocaagil.domain.representations.UsersRepresentation;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.PropertyMap;
