+++ src/main/java/br/com/pipocaagil/apipipocaagil/repositories/UsersRepository.java
@@ -1,6 +1,6 @@
 package br.com.pipocaagil.apipipocaagil.repositories;
 
-import br.com.pipocaagil.apipipocaagil.domain.Users;
+import br.com.pipocaagil.apipipocaagil.domain.entities.Users;
 import br.com.pipocaagil.apipipocaagil.domain.enums.UserPermissionType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
