+++ src/main/java/br/com/pipocaagil/apipipocaagil/services/interfaces/UsersService.java
@@ -1,6 +1,6 @@
 package br.com.pipocaagil.apipipocaagil.services.interfaces;
 
-import br.com.pipocaagil.apipipocaagil.domain.Users;
+import br.com.pipocaagil.apipipocaagil.domain.entities.Users;
 import br.com.pipocaagil.apipipocaagil.domain.enums.UserPermissionType;
 import br.com.pipocaagil.apipipocaagil.domain.representations.UserUpdateRepresentation;
 import br.com.pipocaagil.apipipocaagil.domain.representations.UsersRepresentation;
